feat: add VRBOX/VRBOX2/ALLDIE actors + fix c_dylink 64-bit easyCreate
- Added d_a_vrbox.cpp, d_a_vrbox2.cpp, d_a_alldie.cpp to REL_SOURCES
- Registered VRBOX, VRBOX2, ALLDIE profiles in pal_profile_list
- Fixed c_dylink.cpp easyCreate ambiguity (sizeof returns unsigned long
  on 64-bit, ambiguous between s32/u32 overloads — cast to u32)

+++ src/c/c_dylink.cpp
@@ -845,7 +845,7 @@ int cCc_Init() {
     JKRSetCurrentHeap(prevHeap);
 
     #if DEBUG
-    FixedMemoryCheck* check = FixedMemoryCheck::easyCreate(DMC, sizeof(DMC));
+    FixedMemoryCheck* check = FixedMemoryCheck::easyCreate(DMC, (u32)sizeof(DMC));
     #endif
 
     DMC_initialized = true;
